Added type casting in some places and moved "optional" attribute to be AFTER that __Default*  attribute
svn path=/trunk/mcs/; revision=41869

+++ mcs/class/Microsoft.VisualBasic/Microsoft.VisualBasic/DateAndTime.cs
@@ -213,9 +213,9 @@ namespace Microsoft.VisualBasic
 		
 		public static long DateDiff (DateInterval Interval, 
 			System.DateTime Date1, System.DateTime Date2, 
-			[Optional, __DefaultArgumentValue(FirstDayOfWeek.Sunday)] 
+			[Optional, __DefaultArgumentValue((int)FirstDayOfWeek.Sunday)] 
 			FirstDayOfWeek StartOfWeek, 
-			[Optional, __DefaultArgumentValue(FirstWeekOfYear.Jan1)] 
+			[Optional, __DefaultArgumentValue((int)FirstWeekOfYear.Jan1)] 
 			FirstWeekOfYear StartOfYear) 
 		{
 			
@@ -305,9 +305,9 @@ namespace Microsoft.VisualBasic
 		(
 			Microsoft.VisualBasic.DateInterval Interval, 
 			System.DateTime DateValue, 
-			[Optional, __DefaultArgumentValue(FirstDayOfWeek.Sunday)] 
+			[Optional, __DefaultArgumentValue((int)FirstDayOfWeek.Sunday)] 
 			FirstDayOfWeek StartOfWeek, 
-			[Optional, __DefaultArgumentValue(FirstWeekOfYear.Jan1)] 
+			[Optional, __DefaultArgumentValue((int)FirstWeekOfYear.Jan1)] 
 			FirstWeekOfYear StartOfYear) {
 		
 			CalendarWeekRule WeekRule = CalendarWeekRule.FirstDay;
@@ -385,9 +385,9 @@ namespace Microsoft.VisualBasic
 
 		public static System.Int64 DateDiff (string Interval, 
 			System.Object Date1, System.Object Date2, 
-			[Optional, __DefaultArgumentValue(FirstDayOfWeek.Sunday)]
+			[Optional, __DefaultArgumentValue((int)FirstDayOfWeek.Sunday)]
 			FirstDayOfWeek StartOfWeek, 
-			[Optional, __DefaultArgumentValue(FirstWeekOfYear.Jan1)] 
+			[Optional, __DefaultArgumentValue((int)FirstWeekOfYear.Jan1)] 
 			FirstWeekOfYear StartOfYear) 
 		{
 			if (Date1 == null) {
@@ -410,9 +410,9 @@ namespace Microsoft.VisualBasic
 
 		public static System.Int32 DatePart (string Interval, 
 			System.Object DateValue, 
-			[Optional, __DefaultArgumentValue(FirstDayOfWeek.Sunday)] 
+			[Optional, __DefaultArgumentValue((int)FirstDayOfWeek.Sunday)] 
 			FirstDayOfWeek StartOfWeek, 
-			[Optional, __DefaultArgumentValue(FirstWeekOfYear.Jan1)] 
+			[Optional, __DefaultArgumentValue((int)FirstWeekOfYear.Jan1)] 
 			FirstWeekOfYear StartOfYear) 
 		{
 			if (DateValue == null) {
@@ -548,7 +548,7 @@ namespace Microsoft.VisualBasic
 		}
 
 		public static int Weekday (System.DateTime DateValue, 
-			[Optional, __DefaultArgumentValue(FirstDayOfWeek.Sunday)] 
+			[Optional, __DefaultArgumentValue((int)FirstDayOfWeek.Sunday)] 
 			FirstDayOfWeek StartOfWeek) 
 		{ 
 			return DatePart(DateInterval.Weekday, DateValue, StartOfWeek, FirstWeekOfYear.System);
@@ -570,7 +570,7 @@ namespace Microsoft.VisualBasic
 		
 		public static System.String WeekdayName (int Weekday, 
 			[Optional, __DefaultArgumentValue(false)] System.Boolean Abbreviate, 
-			[Optional, __DefaultArgumentValue(FirstDayOfWeek.System)] 
+			[Optional, __DefaultArgumentValue((int)FirstDayOfWeek.System)] 
 			FirstDayOfWeek FirstDayOfWeekValue) 
 		{ 
 			if (Weekday < 1 || Weekday > 7) {
